When using Checkout As ... if the new project wizard creates anything locally with the exception of the .project file, then you get a warning.  This would include Java projects, since they create a .classpath file.
The warning itself is valid in that the Checkout process will first delete
anything and everything in the target folder.  Presumably this is because
the checkout process would fail if there were existing local resources.

This change simply provides a more specific warning message that explains why
the user is seeing the message.
Issue#: 237