Added support for atomic commits across projects when JavaSVN is the adapter.
JavaSVN added a new commit method that can let us just pass all of the
selections from the commit dialog to it.  It then figures out which files
belong to the same repository and commits them atomically, even if they
are in different projects.  If the files are in projects in different
repositories then it just automatically commits the files in multiple
commits.