If committing a folder delete, always do a recursive commit.
Right now, this means that if in the same commit you also had a folder with
a property-change and modified files beneath that folder.  And you only selected
to commit the folder prop-change, then the files beneath that folder would
also be committed.

I need to figure out the proper way to change the UI to not allow this.
Basically, if a folder delete is selected in the commit and there is also a
folder prop-change selected in the commit, then all files beneath the folder
prop-change also have to be selected, or some warning message should be
displayed.
Issue #: 402