Milestone 1 of new branch/tag support.
Added support for a new subclipse:tags property that contains a list of values
in this format:

revnum,tag_name,relative_path,branch/tag

For example:

123,custom_mods,/branches/custom_mods,branch
145,v1.0,/tags/1.0,tag
188,v1.1,/tags/1.1,tag
225,v2.0,/tags/2.0,tag

These tag values are currently shown everywhere we show history.  We also
support this when showing remote history by querying the repository for the
tag properties.  On slow connections this can be slow, so it is a preference
and off by default.