Fix problem importing a PSF in 3.4 M6+
The lock scope when modifying a project has changed from workspace to project
(which is a good change).  Our PSF import code was only locking the first
project so it would fail when the second project began its checkout.

Modified code to use MultiRule.combine() to create a scheduling rule that
covers the modification of all projects.  I tested this in 3.3, it needs to
be tested in 3.4.