Added an isThreadSafe() method to ISVNClientAdapter.  SVNKit is false, JavaHL is true.
In Subclipse, we no longer cache and reuse the ISVNClientAdapter if it is not threadsafe.  This fixes
bugs when using SVNKit from multiple threads.  Since we will now be doing a lot more calls to the
SVNClientAdapter plugin to get a new ISVNClientAdapter, I changed that plugin to cache all of the
registered adapters.  I do not know that it would be a big performance hit to look them up on each
call, but I assume it must use some resources.

JavaHL adapter will still be cached in SVNClientManager.  If we run into thread safety issues in the
future, we can now just change this boolean in the adapter and we will not have to change Subclipse.