Don't set working copy depth when updating from sync view - follow up
Fixes the problem for new folders as well.  Only do recursive update from common
parent when the selection is NOT a change set.  When a change set is selected,
perform dedicated updates of resources.  For change set updates, unless the
resource is new, update with depth=empty (but do not set working copy depth).
Issue #: 1648