Brute force copy svnclientadapter source to this repository
When using SVN, we were using svn:externals to pull in the
code from another part of the repository. Until we are building
JAR's for all of these pieces that we need, I am just doing
a copy of the code.

This is the version from Subclipse trunk which uses SVN 1.8.x
Will do a followup commit to add the SVN 1.9.x versions.