Option to compare history view revision or changed path on double click
Add a 'Compare to previous on double click' toggle to the history view
dropdown menu.  If option is checked when a revision or changed path is
double clicked, then compare the clicked revision or change path to the
previous revision (with no intermediate dialog) rather than opening it
in the default editor.

Issue #15