Fix CommitCommentArea so that it respects theme colors
Problem was due to this Eclipse bug:

  https://bugs.eclipse.org/bugs/show_bug.cgi?id=539441

Fortunately, the problem only shows itself if the SourceViewer uses a
ruler.  CommitCommentArea applied a ruler to the SourceViewer, but it
was unnecessary (this is not related to the ruler that can be made to
show using the 'tsvn:logwidthmarker' property; that works with or
without applying a ruler to the SourceViewer).

Besides fixing Issue #44 (Commit dialog comment pane) and Issue #45 (Set
commit properties dialog), this fixes many other dialogs and wizards
that make use of CommitCommentArea.

Issue #44
Issue #45