Return gracefully in LocalResource.isSymLink() if resource doesn't exist
This is a follow-up to b76c575781f6e3fbdf4611ba826c7bad5aa0d314, due to
which the sym-link information is stored as a session property. In case
the file doesn't exist (e.g. if it was removed after an SVN update),
this now throws a ResourceException.

In addition to checking whether the resource is null, it should also be
checked whether the resource exists. Given that a non-existing file
can't be a symbolic link, it is sufficient to return early if either of
those conditions are met.